added first functionality for app
user first is choosing folder which they want to copy then file chooser is opened then path is recieved by the function which is needed as source folder. Destination folder is then chosen and path is also recieved. After that copy files should be made. Need to make also some recongnision of servers, Simpy json format document maybe. This cannot be regular copy script only script which copies all of the data to server, after that user  can mark by the checkbox and click a button to delete that data.